fix the bug in the proces,resources,filesystem interface; link:http://172.17.66.192/biz/bug-view-24372.html; link:http://172.17.66.192/biz/bug-view-23838.html; link:http://172.17.66.192/biz/bug-view-23805.html; link:http://172.17.66.192/biz/bug-view-23446.html; link:http://172.17.66.192/biz/bug-view-23796.html

+++ src/processworker.h
@@ -58,10 +58,10 @@ class ProcessWorker
     guint64 cpu_time;
     QString flownet_persec;
     QString diskio_persec;
-    int mNumFlownet;
-    int mNumDiskIo;
+    long long int mNumFlownet;
+    long long int mNumDiskIo;
     guint status;
-    guint pcpu;
+    double pcpu;
     gint nice;
     gchar *tooltip;
     gchar *name;
